Fixed E2M8 problem with track flat.
Work around:

Every new track in GLSpeed uses a huge virtual flat (4096x4096 px) named E2MxFLAT (created at runtime). The flat is cached to disk the first time you run GLSpeed, so it doesn't have to be created each time you run the game.

There is a problem with the E2M8 map, this track was originally designed to be the E1M9 (as a single bonus track to the original game), but later I decided to make a whole new championship (i.e. episode) and the track renamed to E2M8. Unfortunately I forgot to change the track flat, so that's why the problem appears. I couldn't notice the error since the flat was already cached in my PC as E1M9FLAT.